changed back to castle

+++ getRGBValues.m
@@ -6,11 +6,11 @@ cntr = 1;
 
 while cntr <= size(M,2)
     coordinates = round(M(3:4,cntr));
-    RGBvalues(1:3,cntr) = Imf(coordinates(2),coordinates(1),:,2);
+    RGBvalues(1:3,cntr) = double(Imf(coordinates(2),coordinates(1),:,2));
     cntr = cntr + 1;
 end
 
-RGBvalues = RGBvalues./255;
+RGBvalues = RGBvalues./255.0;
 
 end
